refactor! move evaluator core into separate library
The data types and procedures specific for evaluating Emacs Lisp code
have been moved to the `(gypsum elisp-eval environment)` library. The
`(gypsum elisp-eval)` library exports `elisp-eval!` and defines the
important built-in functions, as before.

The reason for refactoring is to allow the built-in functions to be
defined across multiple library which can all import the necessary
data types and procedures required to define a built-in function.
Not all libraries need access to the built-in functions, so it is OK
to keep these separate in the `(gypsum elisp-eval)` library.